Code created: Event, EventQueue, Worker, WorkerManager
Created basic framework for handling Workers and events. All classes
are abstract except for EventQueue.
-------------------------------------------------------------------
Event: Abstract event class. Placeholder right no, no attributes or
methods
-------------------------------------------------------------------
EventQueue: Simple FIFO Event queue. Internally stores Events in an
private ArrayList called queue.
Methods add, size call the corresponding ArrayList methods on queue. Method
pop removes first element of queue.
------------------------------------------------------------------
Worker: abstract worker class. should be overridden to implement
receiver, cook and server.
Has Static variables numWorkers and manager, eventQueue and integer id.
Constructor sets id to current value of numWorkers, then increments it. It then
adds the Worker to manager.
receiveEvent takes an event and adds it to the eventQueue.
processEvent not implemented.
sendEvent takes an event and sends it to the manager.
------------------------------------------------------------------
WorkerManager: abstract manager class. Note that this is an internal class
that handles communication between workers, not a class that should be
accessible to a restaurants manager. WorkerManager should only be
instantiated as a static variable, since it needs to be common to all
Workers to function correctly.